add tabs to stats
